% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeChart.R
\name{gaugeChart}
\alias{gaugeChart}
\alias{.gaugeHeader}
\alias{.gaugeDisplay}
\alias{.gaugeAxis}
\alias{.gaugeBeam}
\title{Gauge Chart}
\usage{
gaugeChart(value, min, max, sectors = gaugeChartSectors(), header, label,
  gp = gpar(), vp, fmt = NULL)

.gaugeHeader(header, vp)

.gaugeDisplay(x, min = 0, max = 1, label = NULL, fmt = NULL,
  gp.beam = gpar(), vp)

.gaugeAxis(min = 0, max = 1, label = NULL, fmt = NULL, r = c(0.6, 1),
  n = 5)

.gaugeBeam(x, r = c(0.6, 1), gp = gpar())
}
\arguments{
\item{value}{numeric, value to display}

\item{min, max}{numeric, minimum and maximum value for gauge chart axis}

\item{sectors}{an object of \link{gaugeChartSectors-class} specifying the
the sectors for gaug chart axis}

\item{header}{string, header of the gauge chart, see header section below}

\item{label}{string, label for the gauge chart axis}

\item{gp}{an object of class \link[grid]{gpar}}

\item{vp}{an \link[grid]{viewport} object defining the plot region for the
gauge chart, see viewports section below}

\item{fmt}{opional format string, if supplied, the \code{value} and axis
ranges \code{min}/\code{max} will be formatted using
\code{\link[base]{sprintf}}}

\item{gp.beam}{an object of class \link[grid]{gpar} specifying the gauge
chart's beam's color}

\item{r}{numeric, inner and outer radius for gauge chart axis}

\item{n}{number of steps for the shading of the axis background}
}
\description{
Plot gauge charts using \link[grid:grid-package]{grid} graphics package.
}
\section{Functions}{
\itemize{
\item \code{.gaugeHeader}: Plot the gauge chart's header.

\item \code{.gaugeDisplay}: the main auxilliary function for plotting gauge
charts.

\item \code{.gaugeAxis}: Plot the gauge chart's axis.

\item \code{.gaugeBeam}: Plot gauge chart's beam.
}}

\section{Viewports}{

The different parts of the gauge chart will be shown in separate
\link[grid]{viewport}s. All \link[grid]{viewport}s inherit from the parent
\link[grid]{viewport} \code{vp}. If \code{vp} is not given, the current
plot region will be taken as paren \link[grid]{viewport}.


The gauge chart will be displayed in a separate viewport with aspect ratio
(i.e. width to height) of 2. The viewport will be fitted in 95% of the
parent viewport's width and height.

The axis labels will be plotted below the viewport with an offset of
\code{\link[grid:unit]{unit(0.5, "lines")}}.
}

\section{Header}{

If \code{header} is given, the gaug chart's \link[grid]{viewport} will be
split in two \link[grid]{viewport}s. The upper one has a height of
\code{\link[grid:unit]{unit(4, "lines")}} and contains the header.
}

\examples{
\dontrun{
sectors <- gaugeChartSectors(200, 100, 0, rev = TRUE)

pushViewport(gp = gpar(fontsize = 9, col = '#d7d7d7', fill = 'white'))

gaugeChart(125, 0, 300, header = 'Tagesanmeldungen',
    sectors = sectors, label = 'on risk')

popViewport()
}

}
\keyword{hplot}
\keyword{internal}
