% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeChartSectors.R
\name{gaugeChartSectors}
\alias{gaugeChartSectors}
\alias{gaugeChartSectors-class}
\alias{.gaugeBeamColor}
\title{Gauge Chart Sectors}
\usage{
gaugeChartSectors(success, warning, danger, colors = c(), rev = FALSE)

.gaugeBeamColor(sectors, x, default = "#808080")
}
\arguments{
\item{success, warning, danger}{numeric, lower and upper limits for KPI ranges
\emph{success}, \emph{warning}, and \emph{danger}, see range definition
section for details}

\item{colors}{color definition for background of KPI ranges; see colors
section for details@details}

\item{rev}{logical, if TRUE, the order of KPI ranges is reverted; see
range definition section for the effects}

\item{sectors}{a \code{gaugeSectors}-class object}

\item{x}{numeric, value of the KPI}

\item{default}{color specification if \code{x} is out side of defined
KPI ranges}
}
\description{
Gaug chart sectors define colored backgrounds for gauge chart axis based
on KPI ranges \emph{success}, \emph{warning}, and \emph{danger}.
}
\section{Functions}{
\itemize{
\item \code{.gaugeBeamColor}: With \code{.gaugeBeamColor} the background
color of the KPI range (\emph{success}, \emph{warning}, or \emph{danger})
given by \code{x} will be returned.
}}

\section{Range Definition}{

The sectors will be ordered according to \emph{success}, \emph{warning},
and \emph{danger}. If \code{rev = TRUE} the order will be reverted.

Range definition \code{success}, \code{warning}, or \code{danger} can be
given either as numeric dupel with lower and upper limit. If they are given
by a single number, this will be assumed to be the lower limit.


The respective KPI range for a given value \code{x} is calculated according
to:
\enumerate{
 \item Select the KPI range with lower bound less or equal to value \code{x}
   and - if specified - the upper bound greater or equal to value \code{x}.
 \item If more than one range is returned with step 1., use the last of the
   KPI ranges (c.f. ordering of KPI ranges above).
}
}

\section{Colors}{

Colors may be specified by a vector of any of the three kinds of \code{R}
color specifications, i.e. either a color name (see
\code{\link[grDevices]{colors}()} for a list of colors), a hexidecimal
string of the form \code{"#rrggbb"} or \code{"#rrggbbaa"} (see
\code{\link[grDevices]{rgb}()}), or a positive integer \code{i} meaning
\code{\link[grDevices]{palette}()[i]}. The colors can either be named
according to the KPI ranges or given using the same sequence as the input
of the KPI ranges.
}

\keyword{classes}
\keyword{internal}
