% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGaugeChart-package.R
\docType{package}
\name{gridGaugeChart}
\alias{gridGaugeChart}
\alias{gridGaugeChart-package}
\alias{gridGaugeChart-package}
\title{Generating Gauge Charts Using Grid}
\description{
The package provides a function to plot simple gauge
charts. It mimics the features of the flexdashboards-package's
gaugeChart function to produce gauge charts for any output
device supported by R's standard graphics facilities.
}
\section{History}{

\describe{
 \item{0.7}{initial beta-version}
}
}

